requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking -o requirements.txt
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
certifi==2024.8.30
    # via requests
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.0
    # via requests
ckanapi==4.8
    # via hdx-python-api
click==8.1.7
    # via typer
defopt==6.4.0
    # via hdx-python-api
dnspython==2.7.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.21.2
    # via defopt
email-validator==2.2.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
frictionless==5.18.0
    # via hdx-python-utilities
hdx-python-api==6.3.5
    # via hdx-scraper-dcc (pyproject.toml)
hdx-python-country==3.8.5
    # via hdx-python-api
hdx-python-utilities==3.7.4
    # via
    #   hdx-scraper-dcc (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
humanize==4.11.0
    # via frictionless
idna==3.10
    # via
    #   email-validator
    #   requests
ijson==3.3.0
    # via hdx-python-utilities
inflect==7.4.0
    # via quantulum3
isodate==0.7.2
    # via frictionless
jinja2==3.1.4
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.23.0
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2024.10.1
    # via jsonschema
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
loguru==0.7.2
    # via hdx-python-utilities
makefun==1.15.6
    # via hdx-python-api
markdown-it-py==3.0.0
    # via rich
marko==2.1.2
    # via frictionless
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via inflect
num2words==0.5.13
    # via quantulum3
openpyxl==3.1.5
    # via hdx-python-utilities
petl==1.7.15
    # via frictionless
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
pydantic==2.10.2
    # via frictionless
pydantic-core==2.27.1
    # via pydantic
pygments==2.18.0
    # via rich
pyphonetics==0.5.3
    # via hdx-python-country
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pyyaml==6.0.2
    # via
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via hdx-python-api
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
requests-file==2.1.0
    # via hdx-python-utilities
rfc3986==2.0.0
    # via frictionless
rich==13.9.4
    # via typer
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
setuptools==75.6.0
    # via ckanapi
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.19.3
    # via ckanapi
six==1.16.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sphinxcontrib-napoleon==0.7
    # via defopt
stringcase==1.2.0
    # via frictionless
structlog==24.4.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.0.0
    # via hdx-python-country
text-unidecode==1.3
    # via python-slugify
typeguard==4.4.1
    # via inflect
typer==0.13.1
    # via frictionless
typing-extensions==4.12.2
    # via
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   typeguard
    #   typer
unidecode==1.3.8
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.2.3
    # via
    #   libhxl
    #   requests
validators==0.34.0
    # via frictionless
wheel==0.45.1
    # via libhxl
xlrd==2.0.1
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.4
    # via hdx-python-utilities
xlsxwriter==3.2.0
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
